+++ src/lib/app/types/streams.ts
@@ -8,6 +8,7 @@ export enum SubscriptionState {
 export enum ChannelType {
     public = 'public',
     private = 'private',
+    privatePlus = 'privatePlus',
 }
 
 export type WriteMessageForm = {