+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.2.7
+
+-   Added input field for unlocking privatePlus channels
+
 ### 0.2.2
 
 -   Add links to identity ids